chore: split downloader into multiple files
